More Routing, Demo data created, small db structure changes, started script endpoints

+++ database/migrations/2018_11_16_091805_create_loadtimes_table.php
@@ -16,7 +16,7 @@ class CreateLoadtimesTable extends Migration
         Schema::create('loadtimes', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
-            $table->decimal('loadtime', 5, 2);
+            $table->integer('loadtime');
             $table->integer('urlID')->unsigned();
             $table->foreign('urlID')->references('id')->on('urls')->onDelete('cascade');
         });
